+++ crates/bevy_asset/src/io/mod.rs
@@ -22,7 +22,7 @@ pub use futures_lite::AsyncWriteExt;
 pub use source::*;
 
 use alloc::sync::Arc;
-use bevy_utils::{BoxedFuture, ConditionalSendFuture};
+use bevy_tasks::{BoxedFuture, ConditionalSendFuture};
 use core::future::Future;
 use core::{
     mem::size_of,
