+++ crates/bevy_asset/src/transformer.rs
@@ -1,6 +1,7 @@
 use crate::{meta::Settings, Asset, ErasedLoadedAsset, Handle, LabeledAsset, UntypedHandle};
 use atomicow::CowArc;
-use bevy_utils::{ConditionalSendFuture, HashMap};
+use bevy_tasks::ConditionalSendFuture;
+use bevy_utils::HashMap;
 use core::{
     borrow::Borrow,
     convert::Infallible,
