+++ crates/bevy_tasks/Cargo.toml
@@ -10,7 +10,9 @@ keywords = ["bevy"]
 
 [features]
 default = ["std", "async_executor"]
+alloc = []
 std = [
+  "alloc",
   "futures-lite/std",
   "async-task/std",
   "spin/std",
